util/cb2replay.sh: Separate interrupt handlers from main init
coreboot can be instructed to output information pertaining to
realmode interrupt handlers installed by the VGA BIOS< and will also
output traces of interrupt execution. It is useful to trace where an
interrupt begins, and where it ends. Adjust the script to separate
interrupt traces from the main initialization trace.

Signed-off-by: Alexandru Gagniuc <mr.nuke.me@gmail.com>