replay/aruba_dp_aux: Return number of recv bytes in do_aux_tran()
This function started out as the first serious RE effort. At that
time, it seemed that what we now know as the 'reply' field was a
status code, and it made sense to return it rather than the number of
incoming bytes. Now we know this field actually represents a byte on
the AUX bus.

It thus makes more sense to not mangle the reply field with possible
error codes. This makes do_aux_tran() symmetrical with the wrappers,
where the return value is either the number of bytes transferred, or
an error code. The reply field is now transferred via a pointer, in
the same manner that the wrappers do it.

Signed-off-by: Alexandru Gagniuc <mr.nuke.me@gmail.com>