radeon_replay: Use getopt() to control what we want to do
As more and more branches popped up, each trying to do something else,
a lot of "#if 0" and similar abominations started appearing. They were
never merged into the main branch, but they did show the need to be
able to select which action or action to perform.

As such, introduce getopt() handling. We support four actions already.

Signed-off-by: Alexandru Gagniuc <mr.nuke.me@gmail.com>