replay:aruba_dp_aux: Fix incorrect register number
The way registered are accessed was changed in commit:
* 3152653 radeon_replay/radeon_dp_aux: Immitate linux context information

However, the offset for register 0x194c was not shifted left, as the
new convention required. As a result, we were actually accessing
register 0x0653. Since we believe this register is used to enable the
AUX pad, also rename it appropriately.

Signed-off-by: Alexandru Gagniuc <mr.nuke.me@gmail.com>