util/aux2c: Refactor script to work with our AUX traces
The script relied on two successive accesses to FIFO ctl register in
order to determine the transaction size. Since our code doesn't have
that redundant write, the script was missing it. Prune out successive
writes to FIFO ctl and use a better heuristic to determine when the
write sets the TX size. The script now works with both atombios traces
and traces of our code.

Signed-off-by: Alexandru Gagniuc <mr.nuke.me@gmail.com>