replay: Rename 'aruba_*' accesors to 'radeon_*'
We want to be able to pass down context information from high-level
functions down to the accessors. This allows us to use the same
accessors in our codebase, as well as linux and coreboot. Since the
use of data that may be contained in such context info is specific
to the codebase, the 'radeon_device' struct is kept opaque for as
long as possible.

Signed-off-by: Alexandru Gagniuc <mr.nuke.me@gmail.com>